feat: Add `SOK_CONFIG` env variable
In production environment it is equal to `/etc/sok/competition.yaml`,
because we have to have the config in a separate directory to mount it
to the host and persist it.

In development, it can be anything else, for example `competition.yaml`.
This way it will load the configuration, from the working directory.